Issue #89 - Rollback unnecessary changes - move similar() to EqualEvaluator
