+++ cql-engine/src/main/java/org/opencds/cqf/cql/runtime/Time.java
@@ -1,13 +1,9 @@
 package org.opencds.cqf.cql.runtime;
 
 import org.joda.time.*;
-import org.joda.time.DateTime;
 
-import java.math.BigDecimal;
-import java.time.Month;
-import java.time.MonthDay;
-import java.time.format.DateTimeFormatter;
-import java.util.Arrays;
+import org.opencds.cqf.cql.elm.execution.EqualEvaluator;
+
 import java.util.TimeZone;
 
 /**
@@ -87,14 +83,14 @@ public class Time extends BaseTemporal {
     }
 
     public Boolean equal(Time other) {
-        return this.similar(other, Value.SimilarityMode.EQUAL);
+        return this.similar(other, EqualEvaluator.SimilarityMode.EQUAL);
     }
 
-    public Boolean similar(Time other, Value.SimilarityMode mode) {
+    public Boolean similar(Time other, EqualEvaluator.SimilarityMode mode) {
         if (this.getPartial().size() != other.getPartial().size()) { // Uncertainty
             return null;
         }
-        if (mode.equals(Value.SimilarityMode.EQUAL)) {
+        if (mode.equals(EqualEvaluator.SimilarityMode.EQUAL)) {
             return other.getJodaDateTime().toInstant().compareTo(this.getJodaDateTime().toInstant()) == 0;
         }
         else {
